modified example and go
