--- tests/test_s3_origin.go
@@ -1,27 +1,0 @@
-package test
-
-import (
-	"testing"
-
-	"github.com/gruntwork-io/terratest/modules/terraform"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestTerraformExample(t *testing.T) {
-	// Arrange
-	terraformOptions := &terraform.Options{
-		TerraformDir: "../examples/s3-origin/.",
-	}
-	defer terraform.Destroy(t, terraformOptions)
-
-	// Act
-	terraform.InitAndApply(t, terraformOptions)
-
-	// Assert
-	assert := assert.New(t)
-
-	
-	outputValue := terraform.Output(t, terraformOptions, "acm_certificate_arn")
-
-	assert.NotNil(outputValue)
-}
